Add ability to add process flow node and return graph Clickup: #2e414j[closed]
- Also modified some database primary keys to BIGSERIAL and changed the
corresponding types in Go to int64

+++ devops/database/sql/V006__ProcessFlowNodes.sql
@@ -5,21 +5,21 @@ CREATE TABLE process_flow_node_types (
 );
 
 CREATE TABLE process_flow_nodes (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     name VARCHAR(256),
-    process_flow_id INTEGER NOT NULL REFERENCES process_flows(id) ON DELETE RESTRICT,
+    process_flow_id BIGINT NOT NULL REFERENCES process_flows(id) ON DELETE RESTRICT,
     description TEXT,
     node_type INTEGER NOT NULL REFERENCES process_flow_node_types(id) ON DELETE RESTRICT
 );
 
 CREATE TABLE process_flow_node_inputs (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     name VARCHAR(256) NOT NULL,
-    parent_node_id INTEGER NOT NULL REFERENCES process_flow_nodes(id)
+    parent_node_id BIGINT NOT NULL REFERENCES process_flow_nodes(id)
 );
 
 CREATE TABLE process_flow_node_outputs (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     name VARCHAR(256) NOT NULL,
-    parent_node_id INTEGER NOT NULL REFERENCES process_flow_nodes(id)
+    parent_node_id BIGINT NOT NULL REFERENCES process_flow_nodes(id)
 );
